# common-docs/openapi/openapi.v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aini-Inu API",
        "description": "Aini-Inu \ubc31\uc5d4\ub4dc REST API \ubb38\uc11c",
        "contact": {
            "name": "Aini-Inu Backend Team"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/api/v1/images/presigned-upload/{token}": {
            "put": {
                "tags": [
                    "Upload"
                ],
                "summary": "Presigned \uc5c5\ub85c\ub4dc \uc218\ud589",
                "description": "\ubc1c\uae09\ubc1b\uc740 token\uc73c\ub85c \ubc14\uc774\ub108\ub9ac \uc774\ubbf8\uc9c0 \uc5c5\ub85c\ub4dc\ub97c \uc218\ud589\ud569\ub2c8\ub2e4.",
                "operationId": "uploadByPresignedUrl",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string",
                                "format": "byte"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/v1/test/auth/token": {
            "post": {
                "tags": [
                    "Internal"
                ],
                "summary": "\ud14c\uc2a4\ud2b8 \ud1a0\ud070 \ubc1c\uae09",
                "description": "\uac1c\ubc1c\ud658\uacbd\uc5d0\uc11c memberId \uae30\uc900 Access/Refresh Token\uc744 \ubc1c\uae09\ud569\ub2c8\ub2e4.",
                "operationId": "generateTestToken",
                "parameters": [
                    {
                        "name": "memberId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseMapStringString"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/v1/images/presigned-url": {
            "post": {
                "tags": [
                    "Upload"
                ],
                "summary": "Presigned URL \ubc1c\uae09",
                "description": "\uc774\ubbf8\uc9c0 \uc5c5\ub85c\ub4dc\uc6a9 1\ud68c\uc131 Presigned URL\uc744 \ubc1c\uae09\ud569\ub2c8\ub2e4.",
                "operationId": "createPresignedUrl",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PresignedImageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\ubc1c\uae09 \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePresignedImageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePresignedImageResponse"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "\ud5c8\uc6a9\ub418\uc9c0 \uc54a\uc740 MIME \ud0c0\uc785",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePresignedImageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "\ud1a0\ud070 \uac31\uc2e0",
                "description": "Refresh Token\uc744 \uac80\uc99d\ud574 \uc0c8\ub85c\uc6b4 Access/Refresh Token\uc744 \ubc1c\uae09\ud569\ub2c8\ub2e4.",
                "operationId": "refresh",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TokenRefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\uac31\uc2e0 \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\uc694\uccad \uac80\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc720\ud6a8\ud558\uc9c0 \uc54a\uac70\ub098 \ub9cc\ub8cc\ub41c \ud1a0\ud070",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLoginResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
        "/api/v1/images/local": {
            "get": {
                "tags": [
                    "Upload"
                ],
                "summary": "\ub85c\uceec \uc774\ubbf8\uc9c0 \uc870\ud68c",
                "description": "\uac1c\ubc1c\ud658\uacbd \ub85c\uceec \uc2a4\ud1a0\ub9ac\uc9c0 \uc774\ubbf8\uc9c0\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "getLocalFile",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponseVoid": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "type": "object",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "ApiResponseMapStringString": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "ApiResponseLoginResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/LoginResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "tokenType": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "memberId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "newMember": {
                        "type": "boolean"
                    },
                    "isNewMember": {
                        "type": "boolean"
                    }
                }
            },
...
            "PresignedImageRequest": {
                "type": "object",
                "properties": {
                    "purpose": {
                        "type": "string",
                        "minLength": 1
                    },
                    "fileName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "contentType": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "contentType",
                    "fileName",
                    "purpose"
                ]
            },
            "ApiResponsePresignedImageResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/PresignedImageResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
            "PresignedImageResponse": {
                "type": "object",
                "properties": {
                    "uploadUrl": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "expiresIn": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "maxFileSizeBytes": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "TokenRefreshRequest": {
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "refreshToken"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}